Position 8/8/8/2p3K1/2P2N2/1P4k1/8/6n1 w - - 0 11 showed that a lot more fixes (bigger integer types) are needed.

+++ RubiChess/tbcore.h
@@ -96,7 +96,7 @@ struct TBEntry_piece {
   ubyte has_pawns;
   ubyte enc_type;
   struct PairsData *precomp[2];
-  int factor[2][TBPIECES];
+  uint64 factor[2][TBPIECES];
   ubyte pieces[2][TBPIECES];
   ubyte norm[2][TBPIECES];
 };
@@ -112,7 +112,7 @@ struct TBEntry_pawn {
   ubyte pawns[2];
   struct {
     struct PairsData *precomp[2];
-    int factor[2][TBPIECES];
+    uint64 factor[2][TBPIECES];
     ubyte pieces[2][TBPIECES];
     ubyte norm[2][TBPIECES];
   } file[4];
@@ -128,7 +128,7 @@ struct DTZEntry_piece {
   ubyte has_pawns;
   ubyte enc_type;
   struct PairsData *precomp;
-  int factor[TBPIECES];
+  uint64 factor[TBPIECES];
   ubyte pieces[TBPIECES];
   ubyte norm[TBPIECES];
   ubyte flags; // accurate, mapped, side
@@ -147,7 +147,7 @@ struct DTZEntry_pawn {
   ubyte pawns[2];
   struct {
     struct PairsData *precomp;
-    int factor[TBPIECES];
+    uint64 factor[TBPIECES];
     ubyte pieces[TBPIECES];
     ubyte norm[TBPIECES];
   } file[4];
